+++ admin/partials/book-review-admin-images.php
@@ -1,6 +1,5 @@
 <p>
-  <?php _e( 'Configure the images to use for displaying ratings.',
-    $this->plugin_name ) ?>
+  <?php _e( 'Configure the images to use for displaying ratings.', $this->plugin_name ); ?>
 </p>
 
 <form action="options.php" method="post">
@@ -11,81 +10,56 @@
 
   <table class="form-table">
     <tbody>
-      <!-- Show rating on home page -->
+      <!-- Show rating in excerpts -->
       <tr>
-        <th>
+        <th scope="row">
           <label for="book_review_rating_home">
-            <?php _e( 'Show rating on home page', $this->plugin_name ) ?>:
-            &nbsp;&nbsp;
-            <a href="#" class="tooltip">
-              <?php echo $tooltip ?>
-              <span>
-                <?php _e( 'Whether to show the rating image on your home ' .
-                  'page when summary text is used.', $this->plugin_name ) ?>
-              </span>
-            </a>
+            <?php _e( 'Excerpts', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
           <input id="book_review_rating_home" type="checkbox"
             name="book_review_ratings[book_review_rating_home]" value="1"
-            <?php echo checked( '1', $ratings['book_review_rating_home'],
-              false ); ?>>
+            <?php echo checked( '1', $ratings['book_review_rating_home'], false ); ?>>
+          <?php _e( 'Show the rating in excerpts',
+            $this->plugin_name ); ?>
         </td>
       </tr>
 
       <!-- Use default rating images -->
       <tr>
-        <th>
+        <th scope="row">
           <label for="book_review_rating_default">
-            <?php _e( 'Use default rating images', $this->plugin_name ) ?>:
-            &nbsp;&nbsp;
-            <a href="#" class="tooltip">
-              <?php echo $tooltip ?>
-              <span>
-                <?php _e( 'Whether to use the default rating images or ' .
-                  'your own.', $this->plugin_name ) ?>
-              </span>
-            </a>
+            <?php _e( 'Default Rating Images', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
           <input id="book_review_rating_default" type="checkbox"
             name="book_review_ratings[book_review_rating_default]" value="1"
-            <?php echo checked( '1', $ratings['book_review_rating_default'],
-              false ); ?>
+            <?php echo checked( '1', $ratings['book_review_rating_default'], false ); ?>
             onchange="showRatingImages();">
+          <?php _e( 'Use the default rating images', $this->plugin_name ); ?>
         </td>
       </tr>
+    </tbody>
+  </table>
 
-      <!-- Rating Image URLs -->
-      <tr>
-        <th>
-          <h4>
-            <label>
-              <?php _e( 'Rating Image URLs', $this->plugin_name ) ?>
-              &nbsp;&nbsp;
-              <a href="#" class="tooltip">
-                <?php echo $tooltip ?>
-                <span>
-                  <?php _e( 'To use your own rating images, enter the ' .
-                    'URL of an image for each rating below (1-5).',
-                    $this->plugin_name ) ?>
-                </span>
-              </a>
-            </label>
-          </h4>
-        </th>
-        <td></td>
-      </tr>
+  <!-- Rating Image URLs -->
+  <div class="rating">
+    <h3><?php _e( 'Rating Image URLs', $this->plugin_name ); ?></h3>
+    <p>
+      <?php _e( 'To use your own rating images, enter the URL of an image for each rating below (1-5).',
+        $this->plugin_name ); ?>
+    </p>
+  </div>
 
+  <table class="form-table rating">
+    <tbody>
       <!-- 1-Star Image URL -->
-      <tr class="rating">
-        <th>
+      <tr>
+        <th scope="row">
           <label for="book_review_rating_image1">
-            <?php
-              _e( 'One-star Image URL', $this->plugin_name );
-            ?>:
+            <?php _e( 'One-star Image URL', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
@@ -97,12 +71,10 @@
       </tr>
 
       <!-- 2-Star Image URL -->
-      <tr class="rating">
-        <th>
+      <tr>
+        <th scope="row">
           <label for="book_review_rating_image2">
-            <?php
-              _e( 'Two-star Image URL', $this->plugin_name );
-            ?>:
+            <?php _e( 'Two-star Image URL', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
@@ -114,12 +86,10 @@
       </tr>
 
       <!-- 3-Star Image URL -->
-      <tr class="rating">
-        <th>
+      <tr>
+        <th scope="row">
           <label for="book_review_rating_image3">
-            <?php
-              _e( 'Three-star Image URL', $this->plugin_name );
-            ?>:
+            <?php _e( 'Three-star Image URL', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
@@ -131,12 +101,10 @@
       </tr>
 
       <!-- 4-Star Image URL -->
-      <tr class="rating">
-        <th>
+      <tr>
+        <th scope="row">
           <label for="book_review_rating_image4">
-            <?php
-              _e( 'Four-star Image URL', $this->plugin_name );
-            ?>:
+            <?php _e( 'Four-star Image URL', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>
@@ -148,12 +116,10 @@
       </tr>
 
       <!-- 5-Star Image URL -->
-      <tr class="rating">
-        <th>
+      <tr>
+        <th scope="row">
           <label for="book_review_rating_image5">
-            <?php
-              _e( 'Five-star Image URL', $this->plugin_name );
-            ?>:
+            <?php _e( 'Five-star Image URL', $this->plugin_name ); ?>:
           </label>
         </th>
         <td>